iett_bot/requirements.txt
selenium
python-telegram-bot
requests
BeautifulSoup4
extprint
